% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisbid3.R
\docType{package}
\name{sisbid3}
\alias{sisbid3}
\alias{sisbid3-package}
\title{Illustrates basic package construction}
\description{
This includes just a few functions, nothing fancy.
It may also include data.
}
