% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{times}
\alias{times}
\title{Multiply together two numbers}
\usage{
times(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The product of \code{x} and \code{y}
}
\description{
Multiply together two numbers
}
\examples{
times(1, 1)
times(10, 1)
}
