% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trivial.R
\name{trivial}
\alias{trivial}
\title{A trivial function}
\usage{
trivial(x, y)
}
\arguments{
\item{x, y}{two values to be summed}
}
\value{
the sum x+y
}
\description{
This will simply add a few numbers
together, which we could do otherwise
}
\examples{
trivial(2, 3)
}
